move error stuff to utils dir
