+++ docs/components/Shell/SideNav.js
@@ -15,7 +15,6 @@ const items = [
     links: [
       { href: '/docs/assertions', children: 'Assertions' },
       { href: '/docs/checks', children: 'Checks' },
-      { href: '/docs/errors', children: 'Errors' },
       { href: '/docs/guards', children: 'Guards' },
       { href: '/docs/opaques', children: 'Opaques' },
       { href: '/docs/utils', children: 'Utils' },
