+++ src/types.ts
@@ -14,6 +14,8 @@ export type ObjectEntry<T> = { [K in keyof T]: [K, T[K]] }[keyof T];
 
 export type Nullish = null | undefined;
 
+export type ErrorLike = { message: string };
+
 // Opaque types
 // ------------------------------
